Fix the globetransformatons fix in previous commit (bc306c6f1bc4122bb87fa8bc6f74ba1ba1227d26)
It broke the TimelineTranslation for when keyframes were using GlobeTranslations. The intial position was not set up correctly for these when the code for finding the globe was done in the update.

Bringing back the check in the matrix/position methods solved the problem. Probably the globe is guaranteed to be created when reaching this function