Add GitHub Actions
Adds GitHub Actions jobs to go along with the Jenkins build server.

This covers more broad range of systems. macOS runners with different versions of Xcode, Ubuntu versions with different compilers and generators, Windows with different versions of Visual Studio, and an automated method to have the latest LTS of Qt installed on every build node.