Fix: Tetramesh volume is created from structured grid
The volume is assumed to be in a structured grid looping over i,j,k (x,y,z) as such the tetrahedralizing step uses this grid to construct correctly formed tetrahedrons. The result is a volume that fits the data shape instead of a boundingbox sourrounding the data.
Adds ui sliders for rendering options