Added the abilty to swap blackhole type in run time
Made it possbile to swap between kerr and Schwarzchild rendering via option property during runtime. The coordinate systems are not aligned but its hard to tell which if any of them are correcly alignd within OpenSpace, kerr should correcly align with the Y up -Z forward standard when working with OpenGL. So stars chould be alligned for kerr, but Schwarzchild is not alligned with the stars for sure and neither are probably rendering the milkyway texture in a correct orientation since it dose not account in for the position of the blackhole.